--- cinemapedia/lib/presentation/widgets/delegates/ssearch_movie_delegate.dart
@@ -1,119 +1,0 @@
-import 'package:animate_do/animate_do.dart';
-import 'package:cinemapedia/config/helpers/huma_formats.dart';
-import 'package:cinemapedia/domain/entities/movies.dart';
-import 'package:flutter/material.dart';
-
-typedef SearchMoviesCallback = Future<List<Movie>> Function(String query);
-
-class SearchMovieDelegate extends SearchDelegate<Movie?> {
-  final SearchMoviesCallback searchMovies;
-  SearchMovieDelegate({required this.searchMovies});
-
-  @override
-  String get searchFieldLabel => 'Buscar pelicula';
-
-  @override
-  List<Widget>? buildActions(BuildContext context) {
-    return [
-      FadeIn(
-        animate: query.isNotEmpty,
-        child: IconButton(
-            onPressed: () => query = '', icon: const Icon(Icons.clear_rounded)),
-      )
-    ];
-  }
-
-  @override
-  Widget? buildLeading(BuildContext context) {
-    return IconButton(
-        onPressed: () => close(context, null),
-        icon: const Icon(Icons.arrow_back_ios_new_rounded));
-  }
-
-  @override
-  Widget buildResults(BuildContext context) {
-    return const Text('buildResults');
-  }
-
-  @override
-  Widget buildSuggestions(BuildContext context) {
-    return FutureBuilder(
-      future: searchMovies(query),
-      builder: (context, snapshot) {
-        final movies = snapshot.data ?? [];
-        return ListView.builder(
-          itemCount: movies.length,
-          itemBuilder: (context, index) {
-            final movie = movies[index];
-            return _MovieItem(movie: movie);
-          },
-        );
-      },
-    );
-  }
-}
-
-class _MovieItem extends StatelessWidget {
-  const _MovieItem({
-    required this.movie,
-  });
-
-  final Movie movie;
-
-  @override
-  Widget build(BuildContext context) {
-    final textStyle = Theme.of(context).textTheme;
-    final size = MediaQuery.of(context).size;
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
-      child: Row(
-        children: [
-          SizedBox(
-            width: size.width * 0.2,
-            child: ClipRRect(
-                borderRadius: BorderRadius.circular(10),
-                child: Image.network(
-                  movie.posterPath,
-                  loadingBuilder: (context, child, loadingProgress) =>
-                      FadeIn(child: child),
-                )),
-          ),
-          const SizedBox(
-            width: 10,
-          ),
-          SizedBox(
-            width: size.width * 0.7,
-            child:
-                Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              Text(
-                movie.title,
-                style: textStyle.titleMedium,
-              ),
-              (movie.overview.length > 100)
-                  ? Text(
-                      '${movie.overview.substring(0, 100)}...',
-                    )
-                  : Text(movie.overview),
-              Row(
-                children: [
-                  Icon(
-                    Icons.star_half_rounded,
-                    color: Colors.yellow.shade800,
-                  ),
-                  const SizedBox(
-                    width: 5,
-                  ),
-                  Text(
-                    HumanFormats.number(movie.voteAverage, 1),
-                    style: textStyle.bodyMedium!
-                        .copyWith(color: Colors.yellow.shade900),
-                  )
-                ],
-              )
-            ]),
-          )
-        ],
-      ),
-    );
-  }
-}
