+++ cinemapedia/lib/presentation/widgets/shared/custom_appbar.dart
@@ -1,13 +1,15 @@
+import 'package:cinemapedia/domain/entities/movies.dart';
 import 'package:cinemapedia/presentation/providers/providers.dart';
 import 'package:cinemapedia/presentation/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomAppbar extends ConsumerWidget {
   const CustomAppbar({super.key});
 
   @override
-  Widget build(BuildContext context,WidgetRef ref) {
+  Widget build(BuildContext context, WidgetRef ref) {
     final colors = Theme.of(context).colorScheme;
     final titleStyle = Theme.of(context).textTheme.titleMedium;
 
@@ -18,16 +20,29 @@ class CustomAppbar extends ConsumerWidget {
         child: SizedBox(
           width: double.infinity,
           child: Row(
-            children: [Icon(Icons.movie_outlined, color: colors.primary),
-            const SizedBox(width: 5,),
-            Text('Cinemapedia', style: titleStyle,),
-            const Spacer(),
-            IconButton(onPressed: (){
-              final movieRepository = ref.read(movieRepositoryProvider);
-              showSearch(context: context, delegate: SearchMovieDelegate(
-                searchMovies: movieRepository.searchMovies
-              ));
-            }, icon: const Icon(Icons.search_outlined))
+            children: [
+              Icon(Icons.movie_outlined, color: colors.primary),
+              const SizedBox(
+                width: 5,
+              ),
+              Text(
+                'Cinemapedia',
+                style: titleStyle,
+              ),
+              const Spacer(),
+              IconButton(
+                  onPressed: () {
+                    final movieRepository = ref.read(movieRepositoryProvider);
+                    showSearch<Movie?>(
+                            context: context,
+                            delegate: SearchMovieDelegate(
+                                searchMovies: movieRepository.searchMovies))
+                        .then((movie) {
+                      if (movie == null) return;
+                      context.push('/movie/${movie.id}');
+                    });
+                  },
+                  icon: const Icon(Icons.search_outlined))
             ],
           ),
         ),
